+++ src/main/java/com/xuanli/oepcms/mapper/SchoolEntityMapper.java
@@ -5,13 +5,13 @@ import org.apache.ibatis.annotations.Mapper;
 import com.xuanli.oepcms.entity.SchoolEntity;
 @Mapper
 public interface SchoolEntityMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(SchoolEntity record);
 
     int insertSelective(SchoolEntity record);
 
-    SchoolEntity selectByPrimaryKey(Integer id);
+    SchoolEntity selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(SchoolEntity record);
